requirements.txt
Pillow
pygame
opencv-python
numpy
mediapipe
RPi.GPIO
fastapi
uvicorn